personalProject/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-bootstrap5==23.4
django-embed-video==1.4.9
django-role-permissions==3.2.0
idna==3.6
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
